alternative redundancy factor for point-to-point targets

+++ monad-node-config/src/fullnode_raptorcast.rs
@@ -19,7 +19,7 @@ pub struct FullNodeRaptorCastConfig<P: PubKey> {
 
     #[serde(bound = "P:PubKey")]
     pub full_nodes_prioritized: FullNodeConfig<P>,
-    pub raptor10_fullnode_redundancy_factor: u8, // validator -> full-nodes
+    pub raptor10_fullnode_redundancy_factor: f32, // validator -> full-nodes
 
     // RaptorCastConfigSecondaryPublisher::GroupSchedulingConfig
     pub max_group_size: usize,
